Added in Dryad Provider
Dryad is a data repository with a focus on biological and life sciences that provides professional curatorial services to researchers in order to meet the expectations of scientific journals. In many respects, Dryad can be treated identically to Dataverse and Figshare. Therefore, these two addons were used as the inspiration for the Dryad addon.

The primary difference between Dryad and the aforementioned other repositories is at the data curation stage. Data is submitted to Dryad via the web interface to a staging or review data repository and arranged into a package. A package consists of a set of files and whatever metadata is available to the submitter. Once the submitter is satisfied with their package, the data is put under review and is available to Dryad curators. Curators may push the package back to the submitter for editing. The review-and-edit process may be repeated several times. Once the package passes curation, the package is moved to Dryad's archive.

The archive database and review space in Dryad are treated separately. As an external application, tracking a package from submission to archive is difficult due to no external identifier that can be used to identify the package in both submission and archive. Once in archive, a package must have a DOI issued by Datacite, and a package may have this same DOI at the time of initial submission, but this is not guaranteed as multiple workflows allow for posteriori DOI creation.